client/requirements.txt
inky==1.3.2
Pillow==9.2.0
numpy==1.23.0